Moves taxonomies list and term usage charts to separate components. #483
