+++ src/assets/css/tainacan-reports.css
@@ -16,5 +16,34 @@
       padding: 0; }
       #tainacan-reports-app .columns .column .postbox {
         margin: 0.75rem; }
+  #tainacan-reports-app .postbox {
+    padding: 1.125rem 1.25rem;
+    margin-bottom: 0;
+    height: 100%;
+    min-height: 380px;
+    background-color: var(--tainacan-block-gray0, #f6f6f6); }
+    #tainacan-reports-app .postbox label {
+      font-weight: bold;
+      font-size: 0.875rem; }
+    #tainacan-reports-app .postbox.metadata-distribution-box {
+      margin: 0px 0px 0.75rem 1.5rem !important;
+      overflow-y: auto; }
+    #tainacan-reports-app .postbox .box-header {
+      display: flex;
+      align-items: baseline;
+      justify-content: space-between;
+      flex-wrap: wrap; }
+      #tainacan-reports-app .postbox .box-header .box-header__item {
+        margin-bottom: 10px; }
+  #tainacan-reports-app .graph-mode-switch {
+    display: inline-block; }
+    #tainacan-reports-app .graph-mode-switch button {
+      border: none !important;
+      background: none !important;
+      box-shadow: none !important;
+      padding: 0;
+      cursor: pointer; }
+      #tainacan-reports-app .graph-mode-switch button.current {
+        color: var(--wp-admin-theme-color, #007cba); }
 
 /*# sourceMappingURL=tainacan-reports.css.map */
