+++ src/views/admin/js/store/modules/report/getters.js
@@ -18,6 +18,10 @@ export const getTaxonomiesList = state => {
   return state.taxonomiesList;
 };
 
+export const getIsFetchingTaxonomiesList = state => {
+  return state.isFetchingTaxomiesList;
+};
+
 export const getTaxonomyTerms = state => {
   return state.taxonomyTerms;
 };
