+++ src/views/admin/js/store/modules/report/index.js
@@ -5,6 +5,7 @@ import * as mutations from './mutations';
 const state = {
     summary: {},
     taxonomiesList: {},
+    isFetchingTaxonomiesList: false,
     colletionsList: {},
     taxonomyTerms: {},
     metadata: {},
