+++ src/views/admin/js/store/modules/report/mutations.js
@@ -18,6 +18,10 @@ export const setTaxonomiesList = (state, taxonomiesList) => {
   state.taxonomiesList = taxonomiesList;
 };
 
+export const setIsFetchingTaxonomiesList = (state, isFetchingTaxomiesList) => {
+  return state.isFetchingTaxomiesList = isFetchingTaxomiesList;
+};
+
 export const setTaxonomyTerms = (state, taxonomyTerms) => {
   state.taxonomyTerms = taxonomyTerms;
 };
