+++ src/views/reports/tainacan-reports.scss
@@ -29,4 +29,46 @@
             }
         }
     }
+
+    .postbox {
+        padding: 1.125rem 1.25rem;
+        margin-bottom: 0;
+        height: 100%;
+        min-height: 380px;
+        background-color: var(--tainacan-block-gray0, #f6f6f6);
+    
+        label {
+            font-weight: bold;
+            font-size: 0.875rem;
+        }
+        &.metadata-distribution-box {
+            margin: 0px 0px 0.75rem 1.5rem !important;  
+            overflow-y: auto;
+        }
+    
+        .box-header {
+            display: flex;
+            align-items: baseline;
+            justify-content: space-between;
+            flex-wrap: wrap;
+    
+            .box-header__item {
+                margin-bottom: 10px;
+            }
+        }
+    }
+    .graph-mode-switch {
+        display: inline-block;
+        button {
+            border: none !important;
+            background: none !important;
+            box-shadow: none !important;
+            padding: 0;
+            cursor: pointer;
+    
+            &.current {
+                color: var(--wp-admin-theme-color, #007cba);
+            }
+        }
+    }
 }
